==> src/03_cv_lin_regr.Rmd <==
---
title: ''
output:
  html_document:
    code_folding: show
    toc: yes
editor_options:
  chunk_output_type: console
---

```{r}
library(tidyverse)
theme_set(theme_minimal())

halldorsson_child_counts <- read_rds("../data/halldorsson_child_counts.rds") |> as.data.frame()
halldorsson_child_fract <- read_rds("../data/halldorsson_child_fract.rds") |> as.data.frame()
```

```{r}
fit_model_with_cv <- function(dataset, model_to_fit, folds = 10, repeats) {
  create_folds <- function(dataset, folds) {
    cvFolds <- cut(seq_len(nrow(dataset)), breaks = folds, labels = FALSE)
    cvFolds <- sample(cvFolds)
    
    predicted <- rep(0, nrow(dataset))
    
    for (i in 1:folds) {
      inTest <- which(cvFolds == i)
      training <- dataset[-inTest, ]
      validation <- dataset[inTest, ]
      fit <- lm(model_to_fit, data = training)
      predicted_values <- predict(fit, newdata = validation)
      predicted[inTest] <- predicted_values
    }
    
    cv_rmse <- sqrt(mean((log10(dataset$age_father) - predicted)^2))
    
    return(cv_rmse)
  }
  
  pd <- tibble(run = 1:repeats)

  for (i in pd$run) {
    set.seed(i)
    pd$cv_rmse[i] <- create_folds(dataset, folds)
    cat("Run", i, "\n")
    flush.console()
  }
  
  assign("pd", pd, envir = .GlobalEnv)
  return(pd)
}
```

```{r}
fit_model_with_cv(halldorsson_child_counts, "log10(age_father) ~ C_G + C_T + CpG_TpG + T_C + T_G + C_A + T_A", 10, 5)
fit_model_with_cv(halldorsson_child_fract, "log10(age_father) ~ C_G + C_T + CpG_TpG + T_C + T_G + C_A + T_A", 10, 5)
```

Better CV RMSE results using the counts data

# Fitting full model on counts data
```{r}
linear_counts_log10 <- lm(log10(age_father) ~ C_G + C_T + CpG_TpG + T_C + T_G + C_A + T_A, data = halldorsson_child_counts)

summary(linear_counts_log10)
hist(linear_counts_log10$residuals)

var(linear_counts_log10$residuals)
```

```{r}
library(patchwork)

p1 <- ggplot() +
  geom_histogram(aes(x = linear_counts_log10$residuals), fill = "deepskyblue3", color = "black", binwidth = 0.05) +
  labs(x = "Distribution of Residuals", y = "Frequency")

set.seed(123)
n <- length(linear_counts_log10_age$residuals)
x <- rnorm(n, sd = sqrt(0.009857264))

p2 <- ggplot(data.frame(x = x, y = linear_counts_log10_age$residuals), aes(sample = linear_counts_log10_age$residuals)) +
  geom_qq() +
  geom_abline(intercept = mean(linear_counts_log10_age$residuals), slope = sd(linear_counts_log10_age$residuals)) +
  labs(x = "Normal Distribution Quantiles", y = "Observed Quantiles")

p <- p1 + p2

ggsave("figure_06_resid.png", plot = p, width = 8, height = 3, path = "./../../Figures/")
```
